리팩토링완료 / validator 추가
