+++ src/refactoring/components/admin/coupon/index.tsx
@@ -1,0 +1,23 @@
+import { Coupon } from "../../../../types";
+import AddCoupon from "./AddCoupon";
+import CouponList from "./CouponList";
+
+interface Props {
+  coupons: Coupon[];
+  onCouponAdd: (newCoupon: Coupon) => void;
+}
+
+const CouponAdmin = ({ coupons, onCouponAdd }:Props) => {
+
+    return (
+        <div>
+          <h2 className="text-2xl font-semibold mb-4">쿠폰 관리</h2>
+          <div className="bg-white p-4 rounded shadow">
+            <AddCoupon onCouponAdd={onCouponAdd} />
+            <CouponList coupons={coupons} />
+          </div>
+        </div>
+    )
+}
+
+export default CouponAdmin;
