+++ src/refactoring/components/admin/product/index.tsx
@@ -1,0 +1,22 @@
+import { Product } from "../../../../types";
+import AddProduct from "./AddProduct";
+import ProductList from "./ProductList";
+
+interface Props {
+    products: Product[];
+    onProductUpdate: (updatedProduct: Product) => void;
+    onProductAdd: (newProduct: Product) => void;
+}
+
+const ProductAdmin = ({ onProductAdd, products, onProductUpdate }: Props) => {
+
+    return (
+        <div>
+            <h2 className="text-2xl font-semibold mb-4">상품 관리</h2>
+            <AddProduct onProductAdd={onProductAdd} />
+            <ProductList products={products} onProductUpdate={onProductUpdate} />
+        </div>
+    )
+}
+
+export default ProductAdmin;
