+++ src/refactoring/entities/product/index.ts
@@ -1,0 +1,3 @@
+import { Product } from "../../../types";
+
+export type FieldValidator = (value: unknown) => boolean;
